Update patch set 2
Patch Set 2: Code-Review+1

I'm totally in favour of a library cap as a way to prevent library releases repeatedly breaking stable CI jobs, but surely the intent was to allow the cap to be raised in a controlled fashion (such as with this change) rather than being forever stuck on a stable branch.

According to semver, 0.6.0 is fully backwards compatible with 0.4.0. Additionally, 0.6.0 has minor feature changes which exposes functionality in stable/kilo heat.

Patch-set: 2