Update patch set 2
Patch Set 2:

> Problem is that on master branch you're automatically getting requirements updates from master global-requirements which will sooner or later get you dependency incompatible with stable/kilo global-requirements.

But, that's a packaging/distribution problem, surely?  I mean, is it justified to fork the *entire codebase* so we can maintain one file separately between branches (e.g requirements.txt)?

It that's the main motivation behind this change, I can imagine we'll end up backporting nearly every patch to the stable/kilo heatclient branch, probably only ommitting stuff related to new-for-liberty features.

Would an alternative be to say clients must have requirements.txt files which are compatible with all non EOL stable branches (e.g the lower bound for all requirements is taken from the oldest supported stable branch)?  

This would mean we have to ensure clients maintain compatibility with a wider range of library versions, but that's basically what we have to do already, so it's no extra work for developers (whereas the stable branch idea really is a significant additional overhead).

Patch-set: 2