Update patch set 2
Patch Set 2: Code-Review+1

Sean, we talked over twitter and I explained why I did this really bad thing. It was a mistake but I don't believe we should go punishing the entire project or team for this. Besides, this is the first time this project has given troubles.

I'll revert this aweful hack that, tbh, was just intended to overcome a limitation in the package manager or a lack of knowledge on my side.

Patch-set: 2
Label: Code-Review=+1