Update patch set 1
Abandoned

In Juno and before, tooz was a stackforge project and was managed manually by Julien Danjou.  Since we don't put stackforge projects in projects.txt, we can't add it.  tooz has a stable/juno branch now for things to go on the 0.12.x series, but those requirements will need to be managed manually.

Patch-set: 1
Status: abandoned