Update patch set 1
Patch Set 1:

If the only bug is that 0.6.4 doesn't correctly indicate its dependency on simplegeneric, then we can address that with a release note or by having the distribution packages properly set up that dependency. If there is a problem in the 0.6.4 version of WSME's code, then we can add an exclusion rule like "!=0.6.4" to the requirements.

Patch-set: 1