Update patch set 2
Patch Set 2:

I see Sean asking who wants to use this in the other review. The answer here is that I have a planned change to oslo's loopingcall for this, but that change isn't done yet. If you'd like me to WIP this then Depends-On for the oslo change I'd be fine with that.

Patch-set: 2