Mark as Python2.7 only the dependencies that are.
We shouldn't try to install on 3.x dependencies that are only
supported on < 3.0. I've used =='2.7' rather than <'3.0' because
setuptools support for < is still spotty and we're tracking that down
and we don't support 2.6 anymore AIUI. I can make this an OR clause
with both 2.6 and 2.7 if we still care about 2.6.

Change-Id: I03667e3e76fd8d84b75dcb5b61742a3f13040662
Depends-On: I31adaf35c8d7b72fe3f8c9242cc356fe34d537e8