Add upper-constraints.txt.
This is a first approximation to the transitive union of requirements,
built by generate-constraints with python 2.7 and 3.4. It will be
updated by a periodic job (which we'll add shortly) proposing future
merges to this repository. Its not yet cross-checked for compatibility
with proposed global-requirements.txt changes, that will come later.

Having the file present now allows us to start working on the
consuming side in parallel and has no downsides, since no projects
jobs will be voting based on this until all the pieces are in place
and battle tested.

Change-Id: Ifc04c81fd8d52051a0f09a6fd61b06da035c513f