Update patch set 1
Patch Set 1: -Code-Review

The last I read it, Julian had removed the wheel which fixed glance's gate. The problem is more intricate than you (or others on that bug) present. It isn't simply the fact that pip auto-builds wheels now, it's that it also caches them.

If pip auto-built a wheel for 2.7 that was cached only for 2.7 and another for 3.x that was only cached for 3.x then this wouldn't be an issue. Auto-building the wheel on 3.x would not have this problem since `python setup.py bdist_wheel` is being run under Python 3. That makes that wheel broken for 2.7 but not for other installations on Python 3. Likewise, if pip auto-builds the wheel under 2.7, it breaks the installation for 3.x but not for other 2.7 installations. So it isn't a bad requirements issue, it's a problem with using old hacks for managing dependencies based on Python versioning. The requirements themselves are fine, it's generating a bad wheel that tries to install them on the wrong version of Python

Patch-set: 1
Reviewer: Gerrit User 12000 <12000@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0