Add babel-django plugin for Horizon
Horizon is migrating to babel to allow translation of AngularJS
based pages.

Rather than use a mix of Django's builtin i18n and babel Horizon
will also use the babel-django plugin for legacy i18n.

babel-django is available on the cheeseshop but will require
packaging.

It is in active maintenance (last updated 2015-04-22 for the Django
1.8 release).

License is BSD

Having this package available will simplify translation for
Horizon's new AngularJS based components and mean that Horizon is
more closely in line with other projects' i18n processes.

Change-Id: I06281979b42b72b26d2cdcfbcb12affadd490a4e