Update patch set 1
Patch Set 1:

- Is the library actively maintained?

Yes.  A single maintainer, but he is extremely responsive to bugs/patches and new releases are made regularly.

Fwiw, the author also attended Vancouver summit and is looking to be more actively involved in OpenStack/Neutron community.

- Is the library good code?

I've looked at it fairly closely and: Yes (imo).

- Is the library python 3 compatible?

Yes (and pure python).

- Is the library license compatible?

Versions >= 0.3.6 are dual licensed GPLv2+ and Apache v2.
Before 0.3.6 it was GPLv2+ only.

- Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

0.3.5 is packaged in Debian/Ubuntu, but we would need a newer version.

0.3.4 is in Fedora, again we'd need a newer version.

- Is the function of this library already covered by other libraries in ``global-requirements.txt``?

No.  There are a few "version 0.1" alternative netlink libraries on pypi, but pyroute2 seems the clear standout in terms of features and current activity.

Patch-set: 1