Update patch set 1
Patch Set 1: Code-Review-1

I'm +2 on adding the library, but we need a minimum version specified. It would also be good to have one or two projects who need the library mentioned in the commit message, so we can keep up with the reason for adding it in our history.

Patch-set: 1
Label: Code-Review=-1