Update patch set 3
Patch Set 3: Code-Review-2

Yes, with Kyle's confirmation the Neutron team does not want to regress from having Python 3 support, and so does not want to add new dependencies to the project that don't support Python 3. When there is a version of the library that supports Python 3, we can add that as the minimum supported version.

Patch-set: 3
Label: Code-Review=-2