Update patch set 3
Patch Set 3: -Code-Review

I would like to honor the cross-project decision of supporting python 3. We are in a dilemma of choosing another library with less features for new neutron features or waiting a library release for new features.
If ryu is the best library, of course the best thing is for the neutron guys interested in some new features to join/encourage ryu development for early python 3 support. It is the open-source way.

Patch-set: 3