Remove hacking and hacking-owned dependencies.
hacking and the cluster of tools it brings in are updated by projects
in a lazy fashion. As such, tracking the things that hacking specifies
inside global-requirements doesn't make sense: we can't compile them
into the upper-constraints file, since repositories do not upgrade
their codebases in lockstep. We already have a special case where
hacking is passed through as-is without checking its presence in
global-requirements.txt, so this is a very simple change.

Change-Id: I9f8dc9d44c347bc3dfda92ef1a2cfb20c1fe8f8c