Update patch set 34
Patch Set 34:

I'm abstaining on this until we get some of the Oslo crowd to review. AFAIK they're the current hold-outs for testing master branches of some repos on 2.6. That could arguably stop now that they have stable branches, but that's a conversation which I think still needs to be initiated.

Patch-set: 34
Label: Verified=0