Cap oslo.il8n to allow for transition away from oslo namespace
This commit adds a version cap to prevent pulling in oslo.il8n v2.0.0.
This latest release removed the namespace package which breaks
glance_store and a bunch of the clients usage of oslo.il8n. We need to
stop using this release (and all future versions of oslo.il8n) until
all the packages we depend on using the oslo namespace for the il8n
module are fixed.

Change-Id: I775fe85bb320a819e32e33b7097332dd75b4aa98