Add oslo.cache and oslo.reports to global requirements
Allow projects to kick the tires of oslo.cache and oslo.reports.
oslo.cache was extracted from keystone and is based on dogpile
interfaces. oslo.reports was extractred from oslo-incubator and
is widely used (copied). We would like projects to try to adopt
these libraries and provide valuable feedback on the API.

Change-Id: I8184ade8989349ffca9a71c4ae1b56ca12fecbec