Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Looks mostly good - but optparse is deprecated ... which has just let me to notice that there's other optparse in the tree. I'm guessing it's a holdover from original update.py, which is fine. But - there's also argparse in the tree, so I don't think we need to be consistent with the old stuff.

I know this is a nit - but we really shouldn't be producing new optparse code.

Patch-set: 1