Add pypowervm to global requirements
The API to interact with PowerVM is the open-source pypowervm library.
Multiple projects require use of the API as PowerVM hypervisor support
is being added to various projects.

The impacted projects are:
- Nova: StackForge project initially.  Proposal for core integration in
  an upcoming release.
- Neutron: Decomposed ML2 Agent
- Ceilometer: Core support

This change adds pypowervm to the global-requirements so that the other
projects can make use of the API.

The version of pypowervm being added supports both Python 2.7 and 3.4.

The API can be found at https://github.com/pypowervm/pypowervm

Change-Id: I6174b613ed12228315c85608c95a1df8dea7a714