Update patch set 1
Patch Set 1:

Thanks for the feedback everyone!

We have validated the support for Python 3.4.  This included getting the CI for pypowervm to run against Python 3.4.  Documentation of pypowervm was updated to indicate this (and the version number incremented).  There is also a tag for the version out in GitHub now.

I will push another patch of this that matches the new version from pypowervm.  I think this takes into account the feedback so far.

Patch-set: 1
Label: Verified=0