Update patch set 2
Patch Set 2:

Thanks for updating the python 3.4 support, Drew.

The only other thing that concerns me is the declaration that the current version is unstable. It can be rather difficult to deal with upgrades across unstable API changes in the libraries. Do you have a time-frame for when you think the API will be stable?

Keep in mind that you don't have to list a lib in the global requirements for development happening in a repository that doesn't sync from the global requirements. It might be a good idea for you to develop some of the drivers you have planned in their own repos until the pypowervm API stabilizes enough that the consuming projects are confident that they can add a dependency to it without causing deployment issues.

Patch-set: 2
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0