Update patch set 2
Patch Set 2: Code-Review-1

As well as the fact that lifeless' work should supersede the requirement for this, I'm -1 because the commit message does not accurately describe the change.

To me at least, "Pin at current release" implies a constraing of "==X.Y.Z". I'd describe this change as "Set max version to 1.0.0" (which might well be the current release and constraing future upgrades - but it won't affect a system with an older release, so in my mind it's not a pin)

Patch-set: 2
Reviewer: Gerrit User 9453 <9453@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1