Update patch set 2
Patch Set 2:

Doug,

You ask a good question and I think it's an important one.  pypowervm is a library that we're open sourcing to simplify integration of PowerVM with OpenStack and other open source projects. In doing so the work has been mainly focused on cleaning up, hardening, and adding/extending features to simplify tasks. The API is quickly approaching the level of stability needed for a 1.0 release, but there is still some ongoing optimization/future development.

We're currently aligning pypowervm development with the development schedule of OpenStack.  The current plan is to mark it version 1.0 around August 21st.  That is a bit before the feature freeze of the Liberty release of OpenStack.

We have been developing our projects in StackForge repos currently.  The nova-powervm and neutron-powervm projects are out in StackForge today.  However, the Ceilometer architecture requires our inspector to be in main tree.  Down the road we also would like to bring Nova into main tree, though that will be after the Liberty release.

I am hoping that by having pypowervm move to 1.0 before Feature Complete of the Liberty release, that helps ease concerns.  If not, are there any other actions that could be done to resolve the concerns?

Thank you!

Patch-set: 2
Reviewer: Gerrit User 8190 <8190@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0