Update patch set 2
Patch Set 2: Code-Review-1

Drew and I chatted about placing that ceilometer plugin in a separate package and loading it through setuptools, as the other plugins are loaded. I believe he's going to explore that approach, for now, and raise this again when the pypowervm lib is stable and the drivers are ready to move in-tree.

Patch-set: 2
Label: Code-Review=-1