WIP: Add capability to run tests in environment installed with constraints
This allows tox to be called in a way that it creates an environment
where pip installs packages with '-c upper-constraints.txt' while not
yet adding it to the default run list.

To test the constrained environment run $(tox -epy27-constraints)

As constraints is only in the latest version of pip, 1.7.0, a shim is
run to ensure the version of pip installed in the virtualenv is new
enough.

Change-Id: Ie643864d2b9c64e30e3d56ce485a44ef34c98392