Fix version conflict between PBR and requirements
Barbican, keystonemiddleware, and perhaps other components have recently
experience build breakages.  The problem is that recently in requirements
there have been additions that use the ";" operator to add requirements
depending on Python versions.  For example:

   python-ldap>=2.4;python_version=='2.7'

The error message looks like:

   error in setup command: 'install_requires' must be a string or list
   of strings containing valid project/version requirement specifiers

Support for this format was added to PBR in version 1.2.  Therefore,
the new minimum support level for PBR now needs to be 1.2.

Change-Id: I608bcbb01c28a3c086e7f91818a5d6e5d8cb5043