Update patch set 2
Abandoned

Netted out with the Ceilometer cores.  We're attempting the separate project approach (which removes the need for the global-requirements change).  The new project is under stackforge named ceilometer-powervm.

We may propose this again at a future time as the PowerVM integration (compute, network, telemetry, etc...) continues.

Thanks everyone for the feedback!

Patch-set: 2
Status: abandoned