Update patch set 1
Patch Set 1: Code-Review-1

Let's wait until this Friday, when we move anchor to the openstack namespace:
https://review.openstack.org/#/c/197673/

Also, there's no check-requirements template used for anchor. Please provide a change on top of 197673 that adds the template - this is part of the "contract". You could make this change "Depends-On" the check-requirement one.

Marking -1 since I suggest to wait for anchor to move to openstack first instead of adding a rename change on Friday and waiting for the project-config change as well.

Patch-set: 1
Label: Code-Review=-1