Update patch set 1
Patch Set 1:

Doug, if you recall a few weeks ago there was the debacle with the screwed up exception handling. The version 0.15.0 solves that issue and one set of exceptions is used across the board. In actual fact Nova is currently using this version and unit tests, CI etc are working properly.
Having the consolidated exception handling is far healthier for the library and will enable us to create a 1.0 version.
So hopefully this can be a healthy step forwards.
In addition to this we have addin two additional Exceptions that need for features that we are adding the Nova and Cinder.

Patch-set: 1
Reviewer: Gerrit User 1653 <1653@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0