Add dnspython3
dnspython3 is made by the same author than dnspython, it comes from the
same website. It's probably the same code base, but compiled to Python
3. It has the same license than dnspython which looks less restrictive
than the MIT license: http://www.dnspython.org/LICENSE.

It looks like dnspython3 is a drop-in replacement for dnspython. Swift
needs dnspython3 to work on Python 3.

dnspython3 doesn't work on Python 2: use an environment marker to only
install it on Python 3.

Change-Id: I93d1e2619e77294ca299c026d57f22f0d9c7cc76
Note: dnspython is already port of global requirements.