Update patch set 1
Patch Set 1:

I think review 198790 is a similar solution to perhaps a different problem.

Bug 198790 is about using a syntax to install a package if a given python version.  For that, it looked like the upper-constraints could be used.  Here, the bug is using it to only install a given package if on a win32 platform.

The same syntax within the requirements is used for both.

But that aside, I think the root issue is perhaps that this syntax (;xxx=='yyy') appears to only be supported at 1.0.0 of pbr (http://git.openstack.org/cgit/openstack-dev/pbr/commit/?id=9026f3627b5ba622ab1c9d6e227a6233d02645f4).  Yet it is used in many projects.  It is used in the global-requirements of openstack-infra.

Examples include:
- pywin32;sys_platform=='win32'
- qpid-python;python_version=='2.7'
- wmi;sys_platform=='win32'
- requests-kerberos>=0.6;python_version=='2.7' # MIT
- MySQL-python;python_version=='2.7'
- pysqlite;python_version=='2.7'
- python-ldap>=2.4;python_version=='2.7'
- enum34;python_version=='2.7' or python_version=='2.6'
- futures>=3.0;python_version=='2.7' or python_version=='2.6'
- happybase>=0.5,!=0.7;python_version=='2.7'
- etc...

I'm not really sure why this works for some but not others.  In my env, I do have pbr .11, and its passing.  Perhaps, since I already had the dependencies installed, it didn't hit the issue?  The bug report shows that others are not able to stack or tox.

I believe what the bug report is stating is that the requirements we have now (in neutron and the global-requirements) will not work with fresh (?) installs if they have start with pbr .11.  Therefore if we need to stick to pbr .11 (based on the comment in 198790 it seems we do) then is a scrub needed on the global-requirements?

I think it may be best to jump into openstack-infra IRC to have a discussion.

Patch-set: 1