Update patch set 2
Abandoned

I'm just going to drop this for now.  We have a workaround for the actual issue that was hit, and at this point I'm not sure the current state of all of the libraries involved here w.r.t their place in juno - and doing the wrong thing is probably as bad as whatever issue we already had (and have a workaround for).

Patch-set: 2
Status: abandoned