Update patch set 1
Patch Set 1:

Do we at least know which projects were affected on master and had fixes to their tests?  Then I could check the git history on those projects and find the test fixes for context.  Looks like fixtures 1.3.0 was released a couple of weeks ago so I was hoping to see an example of the failures, like in a gate run.

Patch-set: 1