Update patch set 2
Patch Set 2:

More background: As we evolve the requirements management, we're planning to institute "lower bounds" checking as well as the upper bounds checking we have now. That is, we'll run jobs with the minimum declared versions of libs as well as the most recent known to work. Doing that will be easier to set up if we know what those bounds are in the first place, and so it seems natural to ask someone adding a new library to express those lower bounds, even if it is version 0.0.1, explicitly as the new library is added. Having all of the entries in the global-requirements.txt file in the same format also makes processing them automatically easier. We have some scripts in the repo that deal with the differences already, but future work will be simpler if we can be consistent.

Patch-set: 2
Label: Verified=0