Fix constraint integration test
The test to validate the constraints file against the global requirements
list was looking for any spec from any package to match. Fix it to only
look at the specs for the package from the constraint being checked.

Add some more unit tests to demonstrate the old conditions that were
failing, and to verify the new logic passes.

Add stdout capture to the tests so we can provide more helpful error
messages to folks contributing patches to the different requirement
lists about why the integration test might be failing.

Change-Id: I46fd11031ba8f775fe62406380bf6417684f1f06