Require all requirements to be constrained or blacklisted
Add a test to ensure that all items in global-requirements.txt are also
listed in either upper-constraints.txt or blacklist.txt.

Add a test to ensure that items in blacklist.txt are not in
upper-constraints.txt.

Add a test to ensure the constraints file format is correct, with 3 =
specifying an absolute constraint.

Fix two constraints which were improperly specified with only 2 =.

Change-Id: I277c706d439c943eb35f9b9f741ce9d126298e5f