Update patch set 1
Patch Set 1: Code-Review-1

the function that changed is only used by the tests in keystoneclient, there's no reason we shouldn't be able to use the new oauthlib. the good part is that it won't impact end users right now, since it's just the tests that are borked.

Patch-set: 1
Label: Code-Review=-1