Add Gunicorn
This patch adds Gunicorn to the global dependencies. It's used in Kuryr
to serve a libnetwork remote driver through a Unix domain socket [1].
Originally uWSGI was a candidate but it's released under GPLv2 and
Gunicorn's license is MIT License.

Ubuntu 14.04LTS trusy has a package which version is 17.5 [2] and RHEL6
does 18.0 [3].

[1]: https://review.openstack.org/#/c/204820/
[2]: http://packages.ubuntu.com/trusty/gunicorn
[3]: http://rpms.famillecollet.com/rpmphp/zoom.php?rpm=python-gunicorn

Change-Id: I175893fbffaba6be2089df15b25b7bd31d5d0213
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>