Update patch set 3
Patch Set 3: Code-Review-1

(5 comments)

I think this needs to be more sophisticated. I can appreciate the win32 dependencies being things that we blacklist, but argparse (for instance) can get pulled in and installed on 2.7 [and doing so is necessary for some edge cases, which pip found out the hard way].

I think this patch has uncovered limitations in the constraints process: pinning pip and setuptools is desirable (but they are stripped by 'freeze' so we probably need to change to list instead).

I suggest perhaps having a third file - optionals.txt, which lists things that we accept as being not-pulled-into-constraints-without-blacklisting.

Patch-set: 3
Label: Code-Review=-1