Move data validation out of the unit tests
Having the data validation happening inside the unit tests makes
understanding requirement update failures confusing because the output
is hard to read and does not explain what the error actually is. Move
the checks to their own command and set up a tox env to run it. A
separate patch in project-config will add a new job for the repository.

Change-Id: Icb717b0f36afb6ea29f50bc6935917dddf47fd4c