Update patch set 2
Patch Set 2: Code-Review-1

I spoke to Dims about the fact that this change introduces a new version without allowing any overlap in the supported range of values. He's going to resubmit it as 2 patches, one to remove the cap and update the constraint with another to raise the minimum. That should let us see how the new version works in the unit tests of the consuming projects before we require that it be used.

Patch-set: 2
Label: Code-Review=-1