Update patch set 1
Patch Set 1:

current global reqs have it as 0.9.7 -> < 1.1.0.  These versions of SQLAlchemy all work in Openstack, the issue fixed in 1.0.8 is a stability issue; is it the goal of global-requirements to be conservatively limited to the *best* version of each package, attempting to move past every possible performance or stability-under-load related issue ?  if so, then yes.  

but if that's the case, then what is upper-constraints for?  And why don't we just make all of global-requirements set up to use "package==version" for everything?

Patch-set: 1