Update patch set 1
Patch Set 1: Code-Review+2

It wasn't clear if say there was just one version of SQLA that had a bug we were working around, or if all versions work but some work better than others. In the first case we would want to express which versions don't work. In the latter case it seems safe to just update the constraints and leave the range wider open.

Since it sounds like we have the latter, case, I'll +2 the change.

Patch-set: 1
Label: Code-Review=+2