Update patch set 2
Patch Set 2: Code-Review-2

This is a procedural -2 until the liberty-2 milestone is released: we aim for co-installability during this period, and raising minimum versions across major releases (which includes by definition API breaks) is very likely to prevent that. I'll remove the -2 as soon as the release team announce the milestone.

Patch-set: 2
Label: Code-Review=-2