Update patch set 1
Patch Set 1:

The documentation in https://review.openstack.org/#/c/201893/ may help you understand.

The 'machine' is this repository: we gate the application of new upstream releases through the pins stored in upper-constraints.txt, and there is a job that runs both on changes to this repo, and nightly, to propose updated values.

To require a bugfixed version of a library, edit global-requirements.txt.

Patch-set: 1