Use original package name when updating requirements
For i.e. "posix_ipc", the update-requirements command created an
update to use "posix-ipc" instead of "posix_ipc".

See for example:
        http://git.openstack.org/cgit/openstack/oslo.concurrency/\
        commit/?id=45ff5571380ea4fbd1cab6dca76bd05ebe59b317

For pypi this works but it's more difficult to translate this name into
something useful for system packages.
It's also confusing that the global-requirements.txt has different
requirements than then projects because of the renaming.
So use the pypi name for requirements.

Change-Id: I4126c382bc4d148c5ee97f1bb5bd7375341d903a