Update patch set 1
Patch Set 1:

When updating the requirements for i.e. oslo.concurrency with "update-requirements -v ~/openstack/oslo.concurrency/", I get now a 'posix-ipc' is not in global-requirements.txt' with an exception afterwards. Which is totally expected. So it seems that it's not so easy to revert the posix_ipc -> posix-ipc switch. Any suggestions howto solve that?

Patch-set: 1
Label: Verified=0