Update patch set 1
Patch Set 1:

(1 comment)

The Octavia project needs this library to support the hot plugging of networks into our amphora load balancers and adding the required routing tables.  We proposed a parallel patchset (sorry, missed this one): https://review.openstack.org/#/c/207671/1

We are requesting a newer version of this package and included the upper-constraints file after running the requirements dependency tools.

As you can see by the +1's on the other patchset folks are looking forward to having this available.

Patch-set: 1
Label: Verified=0