Update patch set 2
Patch Set 2:

lifeless: I'm unsure how most of this information will be useful to anyone looking back after the change has been merged, but I have included a snapshot of current distro versions in commit message.

I believe the only remaining question is "is this good code?" and I feel quite uncomfortable being asked to answer such a subjective/inflammatory/libelous question in a public forum and on permanent record.  I have answered that in the review discussion earlier however and won't be repeating it in the commit message.

(Yes, this is an overly defensive response and I apologise in advance.  Both prior forms of commit message were based on recent examples of similar accepted changes from git history, and .. seriously?)

Patch-set: 2