Update patch set 3
Patch Set 3:

So, I understand your frustration. However - the folk that established this repository, in concert with distributors and operators, set some pretty clear benchmarks and the concerns that they had.

I think its reasonable to say that we should change the concerns and push back on some of these things, but for a current change the choice is to follow the consensus decision so far, or halt and wait while the rules are renegotiated.

Patch-set: 3
Label: Verified=0