Add uwsgi to global-requirements
uWSGI is a project that provides a significant toolchain around
application services with a pluggable architecture. uWSGI can stand
in as a replacement for mod_wsgi and other wsgi implementations.

uWSGI is being proposed to global-requirements so that it can be
properly packaged for the disributions and consumed by (at the very
least) Keystone and Zaqar both in devstack (gate) and non-devstack
deployments. uWSGI also will open the path to more easily support
virtual-environments when running services behind Apache.

Change-Id: Iccab5ca0e5a956f7032dbe79ecdfd3640d2118f9