Cap keystoneclient at <2.0.0
Keystoneclient will soon be releasing the 2.x.x branch to the wild,
for Liberty (and the sanity of all involved) we would like to keep
services, clients, etc on the 1.x.x release. The move to keystoneclient
2.x.x will occur in the Mitaka release as this gives a window to
move all dependant servers/services over to the new keystoneauth1
library due to expected compatibility breaks between keystoneclient
1.x.x and 2.x.x (including but not limited to the way session, adapter
and catalog processing happens).

Change-Id: I2e89d1f8b17dd6865616a5c7a06609601f63a0af