Add 'sortedcontainers' requirement
Sorted containers are one of the primary datastructures
that should be included in python but aren't. And at least one
of the taskflow changes would take advantage of such a structure
if it existed. This adds this requirement so that projects can
use the containers it provides as they see fit.

Library @ https://pypi.python.org/pypi/sortedcontainers

- Is the library actively maintained?

Yes, release occurred in june.

- Is the library good code?

It would appear so, has performance metrics, runtime comparisons
and even testimonals.

- Is the library python 3 compatible?

From the docs, yes.

'''Tested on CPython 2.6, 2.7, 3.2, 3.3,
   3.4 and PyPy 2.2+, PyPy3 2.3.1+'''

- Is the library license compatible?

Yes, Apache2.

- Is the library already packaged in the distros we target (Ubuntu
  latest / Fedora latest)?

Appears to be in ubuntu, fedora possibly?

- Is the function of this library already covered by other libraries
  in ``global-requirements.txt``?

Does not appear to be.

Change-Id: Ib3d626dba7f6dd8c438aa563ebd70db04c449da8