Update patch set 2
Patch Set 2:

I've tried to reproduce this with no luck. Was that a devstack run on a clean VM, or was it a devstack run on a previously used VM?

There is a different potential gate buster in the os-client-config 1.6.2 release, but we've addressed that now (and it diden't affect the gate anyway, because of constraints).

Patch-set: 2
Reviewer: Gerrit User 4190 <4190@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0