Cap urllib3 <=1.10.4 so it's synced up with requests 2.7.0
The requests package in pypi has it's own fork of urllib3 which is kept
in sync with snapshots of urllib3.  As of the latest requests package on
pypi (2.7.0), it's synced with urllib3 1.10.4 per [1] so we should also
cap urllib3 in g-r at that same version because it doesn't work with
urllib3 1.11.

This is also important to note since the distros don't handle packaging
requests in the same way.  For example, on Debian the forked urllib3
that's in the requests pip is left intact whereas on Fedora/CentOS/RHEL
the forked urllib3 package is removed so that python-requests just
depends on python-urllib3 from the distro.

So how this works really depends on if you're installing from pypi or a
distro, and if a distro, which distro and how they are handling things,
i.e. python-glanceclient from Fedora is actually going to require
python-urllib3 as a transitive dependency from python-requests, but on
Debian python-glanceclient only requires python-requests and it gets
urllib3 via the internal fork in the pip.

This isn't a problem in the gate since we're testing against Ubuntu and
pypi and have the requests with internal urllib3, but it causes issues
for anyone testing with packages from Fedora/CentOS/RHEL.

Related-Bug: #1476770

Change-Id: Iefeb9bc123f8c65fecd0cba585ecd3eb349b23a6