Add python-ldap-test to global requirements
python-ldap-test provides an in-memory LDAP server that is better
than the pure-python implemented fake-ldap that is used in keystone.
Utilizing a proper LDAP service (based on UnboundID SDK) that is
less weight than full slapd for unit testing will be generally
beneficial for projects that heavily rely on ldap code.

Change-Id: If295ed1783376a997096350f81379284504b2ba6