Update patch set 2
Patch Set 2: Code-Review-1

Please address the questions from README - I know its tedious - we should perhaps ask less - but while we ask them, we should be minimally consistent in thinking about it. Easiest way is to put it in the commit message where any dispute is easy to read, and where we have an audit trail. THanks.

Patch-set: 2
Label: Code-Review=-1