Set cap on python-swiftclient to 2.4.0
In order to close bug 1486576, we need to create a new 2.3.x release of
swiftclient.  Set the upper bound in stable/juno to 2.4.0 (which has already
been released so we have head room to make the fix.

I believe that this is safe to merge now (befoer 2.3.2 is released) as 2.3.1
continue to be installed.

Change-Id: Ida91c781f098900a828f7139554f0223808544db
Partial-Bug: 1486576