Update patch set 4
Patch Set 4:

I'm confused. This isn't needed with the lower patch it depends on. Is this actually about updating to 2.6.0? (In which case the usual thing is to explain why the update is being done).

Patch-set: 4
Reviewer: Gerrit User 4190 <4190@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0