Update patch set 1
Patch Set 1:

I agree that this is the right change, but the commit message should be updated to be a bit more generic about context. This appears to be the original commit message from the ceilometer commit, where it made sense, but here it doesn't make quite as much sense.

Patch-set: 1