Update patch set 1
Patch Set 1:

Copy from ML thread:

Even if this version is fixed for valid_mac, it appears the netaddr authors made the decision to make a backwards incompatible change WRT to the 'broadcast' attribute on IPNetwork objects that have CIDRs of /31 and /32. This means that all future versions of netaddr will be incompatible with Neutron.

I have a fix for master here: https://review.openstack.org/#/c/218723/, but we will need to cap netaddr in global requirements for kilo and juno and then consider back-porting the changes. Additionally, we should probably release a note that upgrading netaddr is disastrous for all released versions of openstack using Neutron.

Patch-set: 1