Update patch set 1
Patch Set 1:

@armax, in master, we are going to fix the issue in neutron (and horizon).

Actually, we just fixed it in kilo neutron with: https://review.openstack.org/#/c/218852/ that we needed to pass grenade for master fix. So the cap is not needed for neutron anymore. But I guess horizon may still be affected. I will leave to horizon folks to decide whether they backport their fix into kilo, or we cap the version, as in this patch.

Patch-set: 1