Bump python-ironicclient version to 0.8.0
Version 0.8.0 has updates for Ironic which needs to be included in the
Ironic Nova driver. Things like native retry mechanism (so it doesn't
need to be implemented in the driver) and the ability to filter which
fields of the API will be returned (to avoid overflowing the logs with
the full resource representation).

Release change: I507d5eb2d206990af7b1ad42a349d45495e0703e

Change-Id: I9b0c8811bf3c83d09e6d65f6bbddb6073e4b66ec