Add mimic to global requirements to use in tests
Mimic is used to mock APIs for testing purposes. The initial consumer
would be python-ironicclient using it for functional testing.

Q: Is the library actively maintained?
A: The library is maintained, including by some Rackspace/Ironic
   contributors.

Q: Is the library good code?
A: Mimic has very good CI and test coverage, and is good code.

Q: Is the library py3 compatible?
A: Mimic was made py3 compatible with the 2.0 release, specifically
   for integration with OpenStack!

Q: Is the library license compatible?
A: Apache2

Q: Is the library already pacakged in distros we target?
A: Mimic is not currently packaged for a distro, however, as it's
   only being added as a test requirement this shouldn't pose a
   hardship.

Q: Is the function of this library already covered by other
   libraries in global-requirements.txt?
A: To my knowledge, there's no other library that supports mimicing
   other APIs, including time stepping and expected failures as
   needed in testing.

Change-Id: Ie5cb36f5b29f049ae8a3ec49beedea640b0e451a