Add certifi to global requirements
Certifi is required by the auto-generated python binding of
kubernetes, which magnum depends on. Recently, magnum merged a patch,
which temprorarily works around the missing of certifi:

https://review.openstack.org/#/c/221376/

We plan to revert the patch above after certifi is added to the
global requirements. In addition, this commit removed the requirement
of python 2.7 in upper constraints, since certifi supports python 3.4
as well [1].

This library satisfies most requirements in README. This library is:
* is actively maintained
* is good code
* is compatible with py3
* is licensed under Mozilla Public License 2.0
* is not packaged in the distros, since it is python specific
* This library has features that are not yet covered under other
  libraries in global requirements

[1] https://github.com/certifi/python-certifi/blob/master/setup.py

Change-Id: Ib3acfdbae02a6a7a2dd656a6a5925ab74d81b0bc