Add certifi to global requirements
Certifi is required by the auto-generated python binding of
kubernetes, which magnum depends on. Recently, magnum merged a patch,
which temprorarily works around the missing of certifi:

https://review.openstack.org/#/c/221376/

We plan to revert the patch above after certifi is added to the
global requirements. In addition, this commit removed the requirement
of python 2.7 in upper constraints, since certifi supports python 3.4
as well [1].

[1] https://github.com/certifi/python-certifi/blob/master/setup.py

Change-Id: Ib3acfdbae02a6a7a2dd656a6a5925ab74d81b0bc