Update patch set 1
Patch Set 1: Code-Review-2

We're in the stable freeze period leading up to the liberty release, and should not add new requirements.

It's not clear from the commit message here whether this is actually new, or if it's just a reflection of changes in the requests package. If it's the latter, we need to clearly update the version settings for requests as well, or to block the version(s) of requests that require the new package. The resolution of that may require more discussion than is going to be convenient on a patch review, so please start a mailing list thread.

Patch-set: 1