Update patch set 1
Patch Set 1:

The failing change that triggered this suggestion is

 "ldappool>=1.0 # MPL " ->   
 "ldappool>=1.0:python_version=='2.7' # MPL "

You can see here that comments were present before and after. The issue is that comments are being passed through to the specifier handling code. We can either change comments to be signalled via e.g. ; or fixup pbr to handle the # better. I think that handling the # better is appropriate, since we already have lots of comments in setup.cfg files, and forcing everyone to churn is unnecessary.

Patch-set: 1