Raise minimum required python-ceilometerclient to 1.2.0
global-requirements is in a bit of a bad situation with
python-ceilometerclient on stable/juno and stable/kilo with version
caps.

On stable/juno, ceilometerclient is capped at <1.0.13 and on stable/kilo
the minimum required ceilometerclient is >=1.0.13.  This leaves no room
to release fixes for python-ceilometerclient on stable/juno.

And python-ceilometerclient 1.0.12 has an uncapped dependency on
oslo.utils which breaks stable/juno jobs.

So ideally what we'd need to do is sync ceilometerclient on stable/juno
with g-r stable/juno and release that as 1.0.15, blacklist 1.0.13 and
1.0.14 on g-r stable/juno.

Then to avoid overlap with g-r stable/kilo, we need to do a stable/kilo
release of python-ceilometerclient as 1.1.1 and make that the new
minimum required version in stable/kilo and cap g-r stable/kilo at <1.2.

And to avoid overlap with master g-r, we would raise the minimum
required version to >=1.2 (done here).

So basically we leave wiggle room for these branches:

1. stable/juno: 1.0.x
2. stable/kilo: 1.1.x
3. master (stable/liberty): 1.2.x and greater (upper-constraints is at
   1.4.0 right now even though 1.5.0 is available).

Partial-Bug: #1494516

Change-Id: Ieb5949f9231d0f8c02fcb720fe1e60b51195ea81